0.0.8.2a - Added remaining database classes plus implemented few functions.
Added all the remaining database related classes as well as created all
the database sets to contain all of the list of the class objects in the
database. Added some functionality to the database controller allowing
to pull all staff members and all jobs in the database for the
presenter.